# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transport/account:
    get:
      operationId: getTransportAccount
      summary: Verified account and current server-managed role; never a demo cookie
      responses:
        "200":
          description: Verified account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemoActor" }
        "401": { description: Verified sign-in required }
        "403": { description: Account suspended or email not verified }
  /transport/staff:
    get:
      operationId: getTransportStaff
      summary: Admin-only staff permissions
      responses:
        "200":
          description: Staff permission records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StaffPermission" }
        "403": { description: Verified Admin required }
  /transport/staff/{userId}:
    put:
      operationId: setTransportStaff
      summary: Admin assigns or revokes a role for an existing verified Clerk user
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffPermissionInput" }
      responses:
        "200":
          description: Saved permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffPermission" }
        "403": { description: Verified Admin required }
        "409": { description: Cannot remove last active Admin }
  /transport/session:
    get:
      operationId: getDemoTransportSession
      summary: Current signed demo-role session (not verified staff authentication)
      responses:
        "200":
          description: Demo actor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemoActor" }
        "401":
          description: No demo role selected
    post:
      operationId: createDemoTransportSession
      summary: Select a role in the explicitly public demo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DemoSessionInput" }
      responses:
        "200":
          description: Signed demo actor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemoActor" }
    delete:
      operationId: deleteDemoTransportSession
      summary: Clear demo role while retaining the anonymous browser identity
      responses:
        "204": { description: Signed out }
...
  /transport/workspace:
    get:
      operationId: getTransportWorkspace
      summary: Shared fleet, inspection history and role-scoped check-ins
      responses:
        "200":
          description: Current server state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransportWorkspace" }
        "401": { description: Demo session required }
        "503": { description: State unavailable; no local lock fallback }
  /transport/vehicles/{vehicleId}/inspection:
    post:
      operationId: startVehicleInspection
      summary: Atomically lock one physical vehicle, staff roles only
      parameters:
        - { name: vehicleId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransportOperation" }
      responses:
        "200":
          description: Inspection started or idempotent replay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InspectionSession" }
        "403": { description: Staff role required }
        "409": { description: Vehicle unavailable or already inspecting }
  /transport/inspections/{inspectionId}/end:
    post:
      operationId: endVehicleInspection
      summary: Finish own inspection; Admin may finish any
      parameters:
        - { name: inspectionId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Inspection finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InspectionSession" }
        "403": { description: Not authorized to end this inspection }
  /transport/check-ins:
    post:
      operationId: createVehicleCheckIn
      summary: Passenger check-in; locked vehicles rejected and attempt counted atomically
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VehicleCheckInInput" }
      responses:
        "200":
          description: Check-in created or idempotent replay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleCheckIn" }
        "423":
          description: Ticket inspection in progress; new check-in not created
        "409": { description: Out of service or existing active journey }
        "403": { description: Passenger role required }
  /transport/check-ins/{checkInId}/checkout:
    post:
      operationId: checkOutVehicle
      summary: Check out own existing journey even while the vehicle is locked
      parameters:
        - { name: checkInId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleCheckIn" }
        "403": { description: Journey ownership required }
...
components:
  schemas:
...
    DemoActor:
      type: object
      required: [id, role]
      properties:
        id: { type: string }
        role: { type: string, enum: [USER, CONDUCTOR, ADMIN] }
    StaffPermissionInput:
      type: object
      additionalProperties: false
      required: [role, disabled]
      properties:
        role: { type: string, enum: [USER, CONDUCTOR, ADMIN] }
        disabled: { type: boolean }
    StaffPermission:
      type: object
      required: [userId, role, disabled, updatedAt]
      properties:
        userId: { type: string }
        role: { type: string, enum: [USER, CONDUCTOR, ADMIN] }
        disabled: { type: boolean }
        updatedAt: { type: string, format: date-time }
    DemoSessionInput:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [USER, CONDUCTOR, ADMIN] }
    VehicleCheckInInput:
      type: object
      required: [vehicleId, requestId]
      properties:
        vehicleId: { type: string, minLength: 1, maxLength: 150 }
        requestId: { type: string, minLength: 8, maxLength: 100 }
    TransportKind:
      type: string
      enum: [TRAM, BUS, TRAIN, METRO, TROLLEYBUS]
...
    TransportVehicle:
      type: object
      required: [id, cityId, lineId, vehicleNumber, status, latitude, longitude, currentStopId, currentStopName]
      properties:
        id: { type: string }
        cityId: { type: string }
        lineId: { type: string }
        vehicleNumber: { type: string }
        status: { type: string, enum: [ACTIVE, CHECK_IN_LOCKED, OUT_OF_SERVICE] }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        currentStopId: { type: string }
        currentStopName: { type: string }
    InspectionSession:
      type: object
      required: [id, cityId, conductorId, vehicleId, lineId, transportType, startedAt, endedAt, status, numberOfPassengersAlreadyCheckedIn, numberOfBlockedCheckInAttempts, requestId]
      properties:
        id: { type: string }
        cityId: { type: string }
        conductorId: { type: string }
        vehicleId: { type: string }
        lineId: { type: string }
        transportType: { $ref: "#/components/schemas/TransportKind" }
        startedAt: { type: string, description: Server UTC timestamp }
        endedAt: { type: string, nullable: true }
        status: { type: string, enum: [ACTIVE, FINISHED] }
        numberOfPassengersAlreadyCheckedIn: { type: integer }
        numberOfBlockedCheckInAttempts: { type: integer }
        requestId: { type: string }
    VehicleCheckIn:
      type: object
      required: [id, userId, vehicleId, cityId, lineId, checkInTime, checkOutTime, status, requestId]
      properties:
        id: { type: string }
        userId: { type: string }
        vehicleId: { type: string }
        cityId: { type: string }
        lineId: { type: string }
        checkInTime: { type: string }
        checkOutTime: { type: string, nullable: true }
        status: { type: string, enum: [ACTIVE, CHECKED_OUT] }
        requestId: { type: string }
    TransportOperation:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string, minLength: 8, maxLength: 100 }
...
    TransportWorkspace:
      type: object
      required: [actor, vehicles, inspections, checkIns, updatedAt]
      properties:
        actor: { $ref: "#/components/schemas/DemoActor" }
        vehicles: { type: array, items: { $ref: "#/components/schemas/TransportVehicle" } }
        inspections: { type: array, items: { $ref: "#/components/schemas/InspectionSession" } }
        checkIns: { type: array, items: { $ref: "#/components/schemas/VehicleCheckIn" } }
        updatedAt: { type: string }
...